# SBCL-specific optimization for plain blob HTTP output.